+++ src/lib/base/String.cpp
@@ -88,7 +88,7 @@ vformat(const char* fmt, va_list args)
     std::vector<size_t> pos;
     std::vector<size_t> width;
     std::vector<size_t> index;
-    size_t maxIndex = 0;
+    long maxIndex = 0;
     for (const char* scan = fmt; *scan != '\0'; ++scan) {
         if (*scan == '%') {
             ++scan;
